最初の動作確認 (loopback以外は動いた)

+++ keepAliveLED.py
@@ -7,32 +7,37 @@ import time
 import os
 import sys
 import pyping
+import netifaces
 
 # LEDが繋がったピン番号の定義
-LED=18
+LED=20
 
 # チェックする間隔
-INTERVAL=10
+INTERVAL=3
 
 # daemon化するか否か
 DAEMON=False
 PID_FILE='/var/run/keepAliveLED.pid'
 
 # pingする先のアドレス (ルータのアドレスを知っているのがベスト)
-HOST='192.168.0.1'
+HOST='192.168.1.1'
 
 # GPIOのピンの設定
 def setup(pin):
+    GPIO.setwarnings(False)
+    GPIO.cleanup()
     # GPIO指定をGPIO番号で行う
     GPIO.setmode(GPIO.BCM)
-    # GPIO21ピンを出力モードに設定
+    # GPIOピンを出力モードに設定
     GPIO.setup(pin, GPIO.OUT)
+    # GPIOピンをLOWに設定
+    GPIO.output(pin, 0)
 
 # ネットワークにつながっているか否かのチェック
 #   ルータへのpingが通るかどうかで判定する
-def networkcheck(host):
-    ret=pyping.ping(host)
-    if (0==ret):
+def networkCheck(host):
+    val = pyping.ping(host)
+    if val.ret_code==0:
         return True
     return False
 
@@ -46,12 +51,14 @@ def setLED(pin,flag):
 
 # なにかのエラーで終了する場合は，PIDファイルを消去，LEDを消して，GPIOの設定をリセット
 def finish(pin,pidFile):
+    print "exception terminate"
     # LEDのピンをLOWに設定
     GPIO.output(pin, 0)
     # GPIO設定をリセット
     GPIO.cleanup()
     # ファイルを消す
-    os.remove(pidFile)
+    if os.path.exists(pidFile):
+        os.remove(pidFile)
     sys.exit()
 
 def loop(pin,interval,host):
@@ -76,6 +83,7 @@ def fork(pin,pidFile,interval,host):
 if __name__ == '__main__':
     setup(LED)
     if DAEMON:
+        print "daemon"
         try:
             fork(LED,PID_FILE,INTERVAL,HOST)
         except:
