+++ loopBackSound.py
@@ -7,9 +7,14 @@ import RPi.GPIO as GPIO
 #import time
 import os
 import sys
+import threading
+
+# ボタンチェックするか否か
+BUTTON_CHECK=True
+#BUTTON_CHECK=False
 
 # ボタンが繋がったピン番号の定義
-BUTTON=20
+BUTTON=21
 
 # ボタンのタイプ
 BUTTON_TYPE=0 # プルダウン式
@@ -23,40 +28,56 @@ DAEMON=False
 PID_FILE='/var/run/homePhone.pid'
 
 # 音声処理のパラメータ
-RATE=44100
+RATE=88200
 CHUNK=1024
+
+#グローバル変数
+pinState=False # ボタンが押されているか否かを示す変数
+#pinState=True # ボタンが押されているか否かを示す変数
+
+def buttonCallBack(self):
+    global stream
+    while True:
+        if (stream.is_active()):
+            input = stream.read(CHUNK)
+            output = stream.write(input)
+        #if ((1==GPIO.input(BUTTON)) and (stream.is_active())):
+        #    input = stream.read(CHUNK)
+        #    output = stream.write(input)
+        #else:
+        #    break
 
 # GPIOのピンの設定
 def setup(pin,audio,rate,chunk):
-    # GPIO指定をGPIO番号で行う
+    # GPIOの初期化
+    GPIO.setwarnings(False)
     GPIO.setmode(GPIO.BCM)
-    # GPIO21ピンを出力モードに設定
-    GPIO.setup(pin, GPIO.IN)
-    return audio.open(	format = pyaudio.paInt16,
-		channels = 1,
-		rate = rate,
-		frames_per_buffer = chunk,
-		input = True,
-		output = True) # inputとoutputを同時にTrueにする
+    GPIO.setup(pin, GPIO.IN, pull_up_down=GPIO.PUD_DOWN)
+    GPIO.add_event_detect(pin, GPIO.RISING)
+    GPIO.add_event_callback(pin, buttonCallBack)
+    # 音声デバイスのオープン
+    return audio.open(   format = pyaudio.paInt16,
+      channels = 1,
+      rate = rate,
+      frames_per_buffer = chunk,
+      input = True,
+      output = True) # inputとoutputを同時にTrueにする
 
 # なにかのエラーで終了する場合は，PIDファイルを消去，GPIOの設定をリセット, オーディオをクローズ
-def finish(pin,pidFile):
+def finish(audio,stream,pin,pidFile):
     # オーディオデバイスをクローズ
     stream.stop_stream()
     stream.close()
-    p.terminate()
+    audio.terminate()
     # GPIO設定をリセット
     GPIO.cleanup()
     # ファイルを消す
-    os.remove(pidFile)
+    if os.path.exists(pidFile):
+        os.remove(pidFile)
     sys.exit()
 
-# ボタンのチェック
-def checkButton(pin):
-    return GPIO.input(pin)
-
 # daemon化する処理
-def fork(pin,pidFile,stream,chunk):
+def fork(check,pin,pidFile,stream,chunk):
     pid = os.fork()
     if pid > 0:
         f = open(pidFile,'w')
@@ -64,18 +85,23 @@ def fork(pin,pidFile,stream,chunk):
         f.close()
         sys.exit()
     if pid == 0:
-        loop(pin,stream,chunk)
+        loop(check,pin,stream,chunk)
 
 # ループの1ラウンド
 def oneRound(stream,chunk):
-    if (stream.is_active()):
-	input = stream.read(CHUNK)
-	output = stream.write(input)
+    return
+    #if (stream.is_active()):
+    #    input = stream.read(CHUNK)
+    #    output = stream.write(input)
 
 # メインのループ
-def loop(pin,stream,chunk):
+def loop(check,pin,stream,chunk):
+    global pinState
     while True:
-        if (1==buttonCheck(pin)):
+        if check:
+            if pinState:
+                oneRound(stream,chunk)
+        else:
             oneRound(stream,chunk)
 
 # メイン
@@ -87,17 +113,13 @@ if __name__ == '__main__':
         finish(BUTTON,PID_FILE)
     if DAEMON:
         try:
-            fork(BUTTON,PID_FILE,stream,CHUNK)
+            fork(BUTTON_CHECK,BUTTON,PID_FILE,stream,CHUNK)
         except:
-            finish(BUTTON,PID_FILE)
+            finish(audio,stream,BUTTON,PID_FILE)
     else:
         try:
-            loop(BUTTON,stream,CHUNK)
+            loop(BUTTON_CHECK,BUTTON,stream,CHUNK)
         except:
-            finish(BUTTON,PID_FILE)
-
-
-
-
+            finish(audio,stream,BUTTON,PID_FILE)
